Fix warnings

+++ app/src/main/java/com/hippo/ehviewer/ui/scene/DownloadsScene.java
@@ -83,7 +83,6 @@ import com.hippo.ehviewer.R;
 import com.hippo.ehviewer.Settings;
 import com.hippo.ehviewer.client.EhCacheKeyFactory;
 import com.hippo.ehviewer.client.EhUtils;
-import com.hippo.ehviewer.client.data.GalleryInfo;
 import com.hippo.ehviewer.dao.DownloadInfo;
 import com.hippo.ehviewer.dao.DownloadLabel;
 import com.hippo.ehviewer.download.DownloadManager;
@@ -117,6 +116,7 @@ import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
 
+@SuppressLint("RtlHardcoded")
 public class DownloadsScene extends ToolbarScene
         implements DownloadManager.DownloadInfoListener,
         FabLayout.OnClickFabListener, FastScroller.OnDragHandlerListener {
@@ -159,6 +159,7 @@ public class DownloadsScene extends ToolbarScene
     private int mType = -1;
 
     private static void deleteFileAsync(UniFile... files) {
+        //noinspection deprecation
         new AsyncTask<UniFile, Void, Void>() {
             @Override
             protected Void doInBackground(UniFile... params) {
@@ -518,63 +519,56 @@ public class DownloadsScene extends ToolbarScene
         }
 
         int id = item.getItemId();
-        switch (id) {
-            case R.id.action_filter: {
-                new MaterialAlertDialogBuilder(requireActivity())
-                        .setSingleChoiceItems(R.array.download_state, mType + 1, (dialog, which) -> {
-                            mType = which - 1;
-                            updateForLabel();
-                            updateView();
-                            dialog.dismiss();
-                        })
-                        .show();
+        if (id == R.id.action_filter) {
+            new MaterialAlertDialogBuilder(requireActivity())
+                    .setSingleChoiceItems(R.array.download_state, mType + 1, (dialog, which) -> {
+                        mType = which - 1;
+                        updateForLabel();
+                        updateView();
+                        dialog.dismiss();
+                    })
+                    .show();
+            return true;
+        } else if (id == R.id.action_start_all) {
+            Intent intent = new Intent(activity, DownloadService.class);
+            intent.setAction(DownloadService.ACTION_START_ALL);
+            ContextCompat.startForegroundService(activity, intent);
+            return true;
+        } else if (id == R.id.action_stop_all) {
+            if (null != mDownloadManager) {
+                mDownloadManager.stopAllDownload();
+            }
+            return true;
+        } else if (id == R.id.action_open_download_labels) {
+            openDrawer(Gravity.RIGHT);
+            return true;
+        } else if (id == R.id.action_reset_reading_progress) {
+            new MaterialAlertDialogBuilder(getContext())
+                    .setMessage(R.string.reset_reading_progress_message)
+                    .setNegativeButton(android.R.string.cancel, null)
+                    .setPositiveButton(android.R.string.ok, (dialog, which) -> {
+                        if (mDownloadManager != null) {
+                            mDownloadManager.resetAllReadingProgress();
+                        }
+                    }).show();
+            return true;
+        } else if (id == R.id.action_start_all_reversed) {
+            List<DownloadInfo> list = mList;
+            if (list == null) {
                 return true;
             }
-            case R.id.action_start_all: {
-                Intent intent = new Intent(activity, DownloadService.class);
-                intent.setAction(DownloadService.ACTION_START_ALL);
-                ContextCompat.startForegroundService(activity, intent);
-                return true;
-            }
-            case R.id.action_stop_all: {
-                if (null != mDownloadManager) {
-                    mDownloadManager.stopAllDownload();
+            LongList gidList = new LongList();
+            for (int i = list.size() - 1; i > -1; i--) {
+                DownloadInfo info = list.get(i);
+                if (info.state != DownloadInfo.STATE_FINISH) {
+                    gidList.add(info.gid);
                 }
-                return true;
             }
-            case R.id.action_open_download_labels: {
-                openDrawer(Gravity.RIGHT);
-                return true;
-            }
-            case R.id.action_reset_reading_progress: {
-                new MaterialAlertDialogBuilder(getContext())
-                        .setMessage(R.string.reset_reading_progress_message)
-                        .setNegativeButton(android.R.string.cancel, null)
-                        .setPositiveButton(android.R.string.ok, (dialog, which) -> {
-                            if (mDownloadManager != null) {
-                                mDownloadManager.resetAllReadingProgress();
-                            }
-                        }).show();
-                return true;
-            }
-            case R.id.action_start_all_reversed: {
-                List<DownloadInfo> list = mList;
-                if (list == null) {
-                    return true;
-                }
-                LongList gidList = new LongList();
-                for (int i = list.size() - 1; i > -1; i--) {
-                    DownloadInfo info = list.get(i);
-                    if (info.state != DownloadInfo.STATE_FINISH) {
-                        gidList.add(info.gid);
-                    }
-                }
-                Intent intent = new Intent(activity, DownloadService.class);
-                intent.setAction(DownloadService.ACTION_START_RANGE);
-                intent.putExtra(DownloadService.KEY_GID_LIST, gidList);
-                ContextCompat.startForegroundService(activity, intent);
-                return true;
-            }
+            Intent intent = new Intent(activity, DownloadService.class);
+            intent.setAction(DownloadService.ACTION_START_RANGE);
+            intent.putExtra(DownloadService.KEY_GID_LIST, gidList);
+            ContextCompat.startForegroundService(activity, intent);
+            return true;
         }
         return false;
     }
@@ -602,44 +596,43 @@ public class DownloadsScene extends ToolbarScene
         toolbar.inflateMenu(R.menu.drawer_download);
         toolbar.setOnMenuItemClickListener(item -> {
             int id = item.getItemId();
-            switch (id) {
-                case R.id.action_add:
-                    EditTextDialogBuilder builder = new EditTextDialogBuilder(context, null, getString(R.string.download_labels));
-                    builder.setTitle(R.string.new_label_title);
-                    builder.setPositiveButton(android.R.string.ok, null);
-                    AlertDialog dialog = builder.show();
-                    new NewLabelDialogHelper(builder, dialog);
+            if (id == R.id.action_add) {
+                EditTextDialogBuilder builder = new EditTextDialogBuilder(context, null, getString(R.string.download_labels));
+                builder.setTitle(R.string.new_label_title);
+                builder.setPositiveButton(android.R.string.ok, null);
+                AlertDialog dialog = builder.show();
+                new NewLabelDialogHelper(builder, dialog);
+                return true;
+            } else if (id == R.id.action_default_download_label) {
+                DownloadManager dm = mDownloadManager;
+                if (null == dm) {
                     return true;
-                case R.id.action_default_download_label:
-                    DownloadManager dm = mDownloadManager;
-                    if (null == dm) {
-                        return true;
-                    }
+                }
 
-                    List<DownloadLabel> list = dm.getLabelList();
-                    final String[] items = new String[list.size() + 2];
-                    items[0] = getString(R.string.let_me_select);
-                    items[1] = getString(R.string.default_download_label_name);
-                    for (int i = 0, n = list.size(); i < n; i++) {
-                        items[i + 2] = list.get(i).getLabel();
-                    }
-                    new MaterialAlertDialogBuilder(context)
-                            .setTitle(R.string.default_download_label)
-                            .setItems(items, (dialog1, which) -> {
-                                if (which == 0) {
-                                    Settings.putHasDefaultDownloadLabel(false);
+                List<DownloadLabel> list = dm.getLabelList();
+                final String[] items = new String[list.size() + 2];
+                items[0] = getString(R.string.let_me_select);
+                items[1] = getString(R.string.default_download_label_name);
+                for (int i = 0, n = list.size(); i < n; i++) {
+                    items[i + 2] = list.get(i).getLabel();
+                }
+                new MaterialAlertDialogBuilder(context)
+                        .setTitle(R.string.default_download_label)
+                        .setItems(items, (dialog1, which) -> {
+                            if (which == 0) {
+                                Settings.putHasDefaultDownloadLabel(false);
+                            } else {
+                                Settings.putHasDefaultDownloadLabel(true);
+                                String label;
+                                if (which == 1) {
+                                    label = null;
                                 } else {
-                                    Settings.putHasDefaultDownloadLabel(true);
-                                    String label;
-                                    if (which == 1) {
-                                        label = null;
-                                    } else {
-                                        label = items[which];
-                                    }
-                                    Settings.putDefaultDownloadLabel(label);
+                                    label = items[which];
                                 }
-                            }).show();
-                    return true;
+                                Settings.putDefaultDownloadLabel(label);
+                            }
+                        }).show();
+                return true;
             }
             return false;
         });
@@ -819,7 +812,7 @@ public class DownloadsScene extends ToolbarScene
                     for (int i = 0, n = labelRawList.size(); i < n; i++) {
                         labelList.add(labelRawList.get(i).getLabel());
                     }
-                    String[] labels = labelList.toArray(new String[labelList.size()]);
+                    String[] labels = labelList.toArray(new String[0]);
 
                     MoveDialogHelper helper = new MoveDialogHelper(labels, downloadInfoList);
 
@@ -1060,29 +1053,25 @@ public class DownloadsScene extends ToolbarScene
                             popupMenu.inflate(R.menu.download_label_option);
                             popupMenu.show();
                             popupMenu.setOnMenuItemClickListener(item -> {
-                                switch (item.getItemId()) {
-                                    case R.id.menu_label_rename: {
-                                        EditTextDialogBuilder builder = new EditTextDialogBuilder(
-                                                context, label, getString(R.string.download_labels));
-                                        builder.setTitle(R.string.rename_label_title);
-                                        builder.setPositiveButton(android.R.string.ok, null);
-                                        AlertDialog dialog = builder.show();
-                                        new RenameLabelDialogHelper(builder, dialog, label);
-                                        break;
-                                    }
-                                    case R.id.menu_label_remove: {
-                                        new MaterialAlertDialogBuilder(requireContext())
-                                                .setTitle(getString(R.string.delete_label_title))
-                                                .setMessage(getString(R.string.delete_label_message, label))
-                                                .setPositiveButton(R.string.delete, (dialog, which) -> {
-                                                    mDownloadManager.deleteLabel(label);
-                                                    mLabels.remove(position);
-                                                    notifyDataSetChanged();
-                                                })
-                                                .setNegativeButton(android.R.string.cancel, null)
-                                                .show();
-                                        break;
-                                    }
+                                int itemId = item.getItemId();
+                                if (itemId == R.id.menu_label_rename) {
+                                    EditTextDialogBuilder builder = new EditTextDialogBuilder(
+                                            context, label, getString(R.string.download_labels));
+                                    builder.setTitle(R.string.rename_label_title);
+                                    builder.setPositiveButton(android.R.string.ok, null);
+                                    AlertDialog dialog = builder.show();
+                                    new RenameLabelDialogHelper(builder, dialog, label);
+                                } else if (itemId == R.id.menu_label_remove) {
+                                    new MaterialAlertDialogBuilder(requireContext())
+                                            .setTitle(getString(R.string.delete_label_title))
+                                            .setMessage(getString(R.string.delete_label_message, label))
+                                            .setPositiveButton(R.string.delete, (dialog, which) -> {
+                                                mDownloadManager.deleteLabel(label);
+                                                mLabels.remove(position);
+                                                notifyDataSetChanged();
+                                            })
+                                            .setNegativeButton(android.R.string.cancel, null)
+                                            .show();
                                 }
                                 return false;
                             });
@@ -1140,44 +1129,6 @@ public class DownloadsScene extends ToolbarScene
         @Override
         public void onItemDragFinished(int fromPosition, int toPosition, boolean result) {
             notifyDataSetChanged();
-        }
-    }
-
-    private class DeleteDialogHelper implements DialogInterface.OnClickListener {
-
-        private final GalleryInfo mGalleryInfo;
-        private final CheckBoxDialogBuilder mBuilder;
-
-        public DeleteDialogHelper(GalleryInfo galleryInfo, CheckBoxDialogBuilder builder) {
-            mGalleryInfo = galleryInfo;
-            mBuilder = builder;
-        }
-
-        @Override
-        public void onClick(DialogInterface dialog, int which) {
-            if (which != DialogInterface.BUTTON_POSITIVE) {
-                return;
-            }
-
-            // Delete
-            if (null != mDownloadManager) {
-                mDownloadManager.deleteDownload(mGalleryInfo.gid);
-            }
-
-            // Delete image files
-            boolean checked = mBuilder.isChecked();
-            Settings.putRemoveImageFiles(checked);
-            if (checked) {
-                // Remove download path
-                EhDB.removeDownloadDirname(mGalleryInfo.gid);
-                // Delete file
-                UniFile file = SpiderDen.getGalleryDownloadDir(mGalleryInfo);
-                deleteFileAsync(file);
-            }
-            if (mLabelAdapter != null) {
-                initLabels();
-                mLabelAdapter.notifyDataSetChanged();
-            }
         }
     }
 
@@ -1346,7 +1297,7 @@ public class DownloadsScene extends ToolbarScene
             mInflater = getLayoutInflater();
             AssertUtils.assertNotNull(mInflater);
 
-            View calculator = mInflater.inflate(R.layout.item_gallery_list_thumb_height, null);
+            @SuppressLint("InflateParams") View calculator = mInflater.inflate(R.layout.item_gallery_list_thumb_height, null);
             ViewUtils.measureView(calculator, 1024, ViewGroup.LayoutParams.WRAP_CONTENT);
             mListThumbHeight = calculator.getMeasuredHeight();
             mListThumbWidth = mListThumbHeight * 2 / 3;
@@ -1386,7 +1337,7 @@ public class DownloadsScene extends ToolbarScene
             holder.rating.setRating(info.rating);
             TextView category = holder.category;
             String newCategoryText = EhUtils.getCategory(info.category);
-            if (!newCategoryText.equals(category.getText())) {
+            if (!newCategoryText.contentEquals(category.getText())) {
                 category.setText(newCategoryText);
                 category.setBackgroundColor(EhUtils.getCategoryColor(info.category));
             }
@@ -1435,7 +1386,7 @@ public class DownloadsScene extends ToolbarScene
         }
     }
 
-    private class ThumbDataContainer implements DataContainer {
+    private static class ThumbDataContainer implements DataContainer {
 
         private final DownloadInfo mInfo;
         @Nullable
